Prevent timeout errors when storing files
According to aws-ruby's maintainer upload timeouts typically occur
because the content length couldn't be determined directly. The fix is
fairly simple, just pass a path to the file instead of the file data and
the gem will open it in binary mode itself.

An integration test (and all of the support around that) had to be put
in place to maintain confidence about the test.