Test for matching ACL grants on copied file
Until now we didn't ensure that the files generated with
CarrierWave::Storage::AWSFile#copy_to were given the same ACL
(Access Control List) as the original files.

As explained in https://github.com/sorentwo/carrierwave-aws/issues/81#issue-154100161
the behavior of #copy_to in CarrierWave is use the `fog_public` configuration
option and make files `public-read` (world-accessible) if it is true.

carrierwave-aws uses a different configuration option (`aws_acl`) which
explicitely defines the default ACL for all uploaded files (i.e. `public-read`)
so it makes sense to compare the copied file with the originally uploaded one.

An alternative test method would be to always check that copied files include
an ACL that matches the `aws_acl` but I cannot for the life of me figure out a
way to get aws-sdk-ruby to output the shorthand `acl` (i.e. `public-read` and
not a list of grants) so for now I'm simply checking that the ACL grants for the
copy are equal to those of the original.