Pass aws_acl to aws-sdk-ruby's #copy_from
This will give the copied file the same ACL grants as originally uploaded
files and prevent the issue demonstrated in #81.

This is admittedly naive since I don't know what happens when aws_acl is nil
or whether it can ever be nil.

CarrierWave::Uploader::Base.normalized_acl suggests that if aws_acl is ever nil
or invalid, a ConfigurationError should be raised so I'm going to hope @sorentwo
is okay with this implementation.

The copying_files_spec is green locally with this change, and since Travis
can't run it, you'll just have to take my word for it. :-p